Restructure noncompliant codes
