+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "vulnerability": "node vulnerability.js"
+    "sonar:vulnerability": "node ./vulnerability"
   },
   "repository": {
     "type": "git",
@@ -17,5 +17,8 @@
   "bugs": {
     "url": "https://github.com/wheelyumdev-test/sonar/issues"
   },
-  "homepage": "https://github.com/wheelyumdev-test/sonar#readme"
+  "homepage": "https://github.com/wheelyumdev-test/sonar#readme",
+  "dependencies": {
+    "request": "^2.88.2"
+  }
 }
