--- vulnerability.js
@@ -1,17 +1,0 @@
-let https = require("https");
-// 17: Server hostnames should be verified during SSL/TLS connections
-// https built-in module:
-let options = {
-  hostname: "www.example.com",
-  port: 443,
-  path: "/",
-  method: "GET",
-  secureProtocol: "TLSv1_2_method",
-  checkServerIdentity: function () {}, // Noncompliant: hostname is not verified
-};
-
-let req = https.request(options, (res) => {
-  res.on("data", (d) => {
-    process.stdout.write(d);
-  });
-}); // Noncompliant
