+++ vulnerability/17/httpsModule.js
@@ -1,0 +1,16 @@
+let https = require("https");
+console.log("https");
+let options = {
+  hostname: "www.example.com",
+  port: 443,
+  path: "/",
+  method: "GET",
+  secureProtocol: "TLSv1_2_method",
+  checkServerIdentity: function() {} // Noncompliant: hostname is not verified
+};
+
+let req = https.request(options, res => {
+  res.on("data", d => {
+    process.stdout.write(d);
+  });
+}); // Noncompliant
