+++ vulnerability/17/requestModule.js
@@ -1,0 +1,7 @@
+let request = require("request");
+console.log("request");
+let socket = request.get({
+  url: "https://www.example.com",
+  secureProtocol: "TLSv1_2_method",
+  checkServerIdentity: function() {} // Noncompliant: hostname is not verified
+});
